Replacement Post350 (#137)
* Complete PrivateProfile intercepts for FRF

* Full PrivateProfile support

* Full PrivateProfile support

* Added PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* Changes for script flexibility: Can be in any folder inside or outside package, may use quotation marks, may use new pseudovariables to reference the package root or user redirection area.

* Updates to documentation on scripting.

* Redo of Post350 Changes

* Update for naming conventions

* Added optional executionpolicy options to script configs.

* Correction to scripting Bypass

* Scripting documentation for ExecutionPolicy

* Fix issue with sending files to StartingScriptWrapper.

* Put wait back in for Shell Launch

* Requested cleanups to PR137

* Document UrlDecode method.

* Cleanup StripFileColonSlash

* wrong merge fix for calling StartProcess

+++ fixups/FileRedirectionFixup/RemoveDirectoryFixup.cpp
@@ -14,23 +14,30 @@ BOOL __stdcall RemoveDirectoryFixup(_In_ const CharT* pathName) noexcept
     {
         if (guard)
         {
-            LogString(L"RemoveDirectoryFixup for pathName", pathName);
+            DWORD RemoveDirectoryInstance = ++g_FileIntceptInstance;
+            LogString(RemoveDirectoryInstance,L"RemoveDirectoryFixup for pathName", pathName);
             
-
-            // NOTE: See commentary in DeleteFileFixup for limitations on deleting files/directories
-            auto [shouldRedirect, redirectPath, shouldReadonly] = ShouldRedirect(pathName, redirect_flags::none);
-            if (shouldRedirect)
+            if (!IsUnderUserAppDataLocalPackages(pathName))
             {
-                if (!impl::PathExists(redirectPath.c_str()) && impl::PathExists(pathName))
+                // NOTE: See commentary in DeleteFileFixup for limitations on deleting files/directories
+                auto [shouldRedirect, redirectPath, shouldReadonly] = ShouldRedirect(pathName, redirect_flags::none);
+                if (shouldRedirect)
                 {
-                    // If the directory does not exist in the redirected location, but does in the non-redirected
-                    // location, then we want to give the "illusion" that the delete succeeded
-                    return TRUE;
+                    if (!impl::PathExists(redirectPath.c_str()) && impl::PathExists(pathName))
+                    {
+                        // If the directory does not exist in the redirected location, but does in the non-redirected
+                        // location, then we want to give the "illusion" that the delete succeeded
+                        return TRUE;
+                    }
+                    else
+                    {
+                        return impl::RemoveDirectory(redirectPath.c_str());
+                    }
                 }
-                else
-                {
-                    return impl::RemoveDirectory(redirectPath.c_str());
-                }
+            }
+            else
+            {
+                Log(L"[%d]Under LocalAppData\\Packages, don't redirect", RemoveDirectoryInstance);
             }
         }
     }
